refactor: create style-utils file to include style transformation common methods
